+++ oci/autoscaling_auto_scaling_configuration_scenario_test.go
@@ -4,6 +4,8 @@
 package oci
 
 import (
+	"fmt"
+	"strconv"
 	"testing"
 	"time"
 
@@ -16,6 +18,36 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	oci_core "github.com/oracle/oci-go-sdk/core"
+)
+
+var (
+	AutoScalingConfigurationResourceConfigForScheduledExecution = AutoScalingConfigurationResourceDependencies +
+		generateResourceFromRepresentationMap("oci_autoscaling_auto_scaling_configuration", "test_auto_scaling_configuration", Optional, Update, autoScalingConfigurationRepresentation2)
+
+	autoScalingConfigurationRepresentation2 = map[string]interface{}{
+		"auto_scaling_resources": RepresentationGroup{Required, autoScalingConfigurationAutoScalingResourcesRepresentation},
+		"compartment_id":         Representation{repType: Required, create: `${var.compartment_id}`},
+		"policies":               RepresentationGroup{Required, autoScalingConfigurationPoliciesRepresentationForScheduledExecution},
+		"cool_down_in_seconds":   Representation{repType: Optional, create: `300`, update: `400`},
+		"defined_tags":           Representation{repType: Optional, create: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "value")}`, update: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "updatedValue")}`},
+		"display_name":           Representation{repType: Optional, create: `example_autoscaling_configuration`, update: `displayName2`},
+		"freeform_tags":          Representation{repType: Optional, create: map[string]string{"Department": "Finance"}, update: map[string]string{"Department": "Accounting"}},
+		"is_enabled":             Representation{repType: Optional, create: `false`, update: `true`},
+	}
+
+	autoScalingConfigurationPoliciesRepresentationForScheduledExecution = map[string]interface{}{
+		"capacity":           RepresentationGroup{Required, autoScalingConfigurationPoliciesCapacityRepresentation},
+		"policy_type":        Representation{repType: Required, create: `scheduled`, update: `scheduled`},
+		"display_name":       Representation{repType: Optional, create: `example_autoscaling_configuration`, update: `displayName2`},
+		"execution_schedule": RepresentationGroup{Optional, autoScalingConfigurationPoliciesExecutionScheduleRepresentation},
+		"is_enabled":         Representation{repType: Optional, create: `true`},
+	}
+
+	autoScalingConfigurationPoliciesExecutionScheduleRepresentation = map[string]interface{}{
+		"expression": Representation{repType: Required, create: `0 15 10 ? * *`},
+		"timezone":   Representation{repType: Required, create: `UTC`},
+		"type":       Representation{repType: Required, create: `cron`},
+	}
 )
 
 type ResourceAutoScalingConfigurationTestSuite struct {
@@ -233,6 +265,234 @@ func (s *ResourceAutoScalingConfigurationTestSuite) TestAccResourceAutoScalingCo
 				Config:             s.Config + tokenFn(TFInstancePool, updatedValues),
 				PlanOnly:           true,
 				ExpectNonEmptyPlan: true,
+			},
+		},
+	})
+}
+
+func TestAutoScalingAutoScalingConfigurationResource_scheduledExecution(t *testing.T) {
+	httpreplay.SetScenario("TestAutoScalingAutoScalingConfigurationResource_basic")
+	defer httpreplay.SaveScenario()
+
+	provider := testAccProvider
+	config := testProviderConfig()
+
+	compartmentId := getEnvSettingWithBlankDefault("compartment_ocid")
+	compartmentIdVariableStr := fmt.Sprintf("variable \"compartment_id\" { default = \"%s\" }\n", compartmentId)
+
+	compartmentIdU := getEnvSettingWithDefault("compartment_id_for_update", compartmentId)
+	compartmentIdUVariableStr := fmt.Sprintf("variable \"compartment_id_for_update\" { default = \"%s\" }\n", compartmentIdU)
+
+	resourceName := "oci_autoscaling_auto_scaling_configuration.test_auto_scaling_configuration"
+	datasourceName := "data.oci_autoscaling_auto_scaling_configurations.test_auto_scaling_configurations"
+	singularDatasourceName := "data.oci_autoscaling_auto_scaling_configuration.test_auto_scaling_configuration"
+
+	var resId, resId2 string
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() { testAccPreCheck(t) },
+		Providers: map[string]terraform.ResourceProvider{
+			"oci": provider,
+		},
+		CheckDestroy: testAccCheckAutoScalingAutoScalingConfigurationDestroy,
+		Steps: []resource.TestStep{
+			// verify create with optionals
+			{
+				Config: config + compartmentIdVariableStr + AutoScalingConfigurationResourceDependencies +
+					generateResourceFromRepresentationMap("oci_autoscaling_auto_scaling_configuration", "test_auto_scaling_configuration", Optional, Create, autoScalingConfigurationRepresentation2),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "auto_scaling_resources.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "auto_scaling_resources.0.id"),
+					resource.TestCheckResourceAttr(resourceName, "auto_scaling_resources.0.type", "instancePool"),
+					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
+					resource.TestCheckResourceAttr(resourceName, "cool_down_in_seconds", "300"),
+					resource.TestCheckResourceAttr(resourceName, "defined_tags.%", "1"),
+					resource.TestCheckResourceAttr(resourceName, "display_name", "example_autoscaling_configuration"),
+					resource.TestCheckResourceAttr(resourceName, "freeform_tags.%", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "id"),
+					resource.TestCheckResourceAttr(resourceName, "is_enabled", "false"),
+					resource.TestCheckResourceAttr(resourceName, "policies.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "policies.0.id"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.initial", "2"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.max", "3"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.min", "2"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.display_name", "example_autoscaling_configuration"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.expression", "0 15 10 ? * *"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.timezone", "UTC"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.type", "cron"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.is_enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.policy_type", "scheduled"),
+					resource.TestCheckResourceAttrSet(resourceName, "policies.0.time_created"),
+					resource.TestCheckResourceAttrSet(resourceName, "time_created"),
+
+					func(s *terraform.State) (err error) {
+						resId, err = fromInstanceState(s, resourceName, "id")
+						if isEnableExportCompartment, _ := strconv.ParseBool(getEnvSettingWithDefault("enable_export_compartment", "false")); isEnableExportCompartment {
+							if errExport := testExportCompartmentWithResourceName(&resId, &compartmentId, resourceName); errExport != nil {
+								return errExport
+							}
+						}
+						return err
+					},
+				),
+			},
+
+			// verify update to the compartment (the compartment will be switched back in the next step)
+			{
+				Config: config + compartmentIdVariableStr + compartmentIdUVariableStr + AutoScalingConfigurationResourceDependencies +
+					generateResourceFromRepresentationMap("oci_autoscaling_auto_scaling_configuration", "test_auto_scaling_configuration", Optional, Create,
+						representationCopyWithNewProperties(autoScalingConfigurationRepresentation2, map[string]interface{}{
+							"compartment_id": Representation{repType: Required, create: `${var.compartment_id_for_update}`},
+						})),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "auto_scaling_resources.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "auto_scaling_resources.0.id"),
+					resource.TestCheckResourceAttr(resourceName, "auto_scaling_resources.0.type", "instancePool"),
+					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentIdU),
+					resource.TestCheckResourceAttr(resourceName, "cool_down_in_seconds", "300"),
+					resource.TestCheckResourceAttr(resourceName, "defined_tags.%", "1"),
+					resource.TestCheckResourceAttr(resourceName, "display_name", "example_autoscaling_configuration"),
+					resource.TestCheckResourceAttr(resourceName, "freeform_tags.%", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "id"),
+					resource.TestCheckResourceAttr(resourceName, "is_enabled", "false"),
+					resource.TestCheckResourceAttr(resourceName, "policies.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "policies.0.id"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.initial", "2"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.max", "3"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.min", "2"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.display_name", "example_autoscaling_configuration"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.expression", "0 15 10 ? * *"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.timezone", "UTC"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.type", "cron"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.is_enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.policy_type", "scheduled"),
+					resource.TestCheckResourceAttrSet(resourceName, "policies.0.time_created"),
+					resource.TestCheckResourceAttrSet(resourceName, "time_created"),
+
+					func(s *terraform.State) (err error) {
+						resId, err = fromInstanceState(s, resourceName, "id")
+						return err
+					},
+				),
+			},
+
+			// verify updates to updatable parameters
+			{
+				Config: config + compartmentIdVariableStr + AutoScalingConfigurationResourceDependencies +
+					generateResourceFromRepresentationMap("oci_autoscaling_auto_scaling_configuration", "test_auto_scaling_configuration", Optional, Update, autoScalingConfigurationRepresentation2),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "auto_scaling_resources.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "auto_scaling_resources.0.id"),
+					resource.TestCheckResourceAttr(resourceName, "auto_scaling_resources.0.type", "instancePool"),
+					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
+					resource.TestCheckResourceAttr(resourceName, "cool_down_in_seconds", "400"),
+					resource.TestCheckResourceAttr(resourceName, "defined_tags.%", "1"),
+					resource.TestCheckResourceAttr(resourceName, "display_name", "displayName2"),
+					resource.TestCheckResourceAttr(resourceName, "freeform_tags.%", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "id"),
+					resource.TestCheckResourceAttr(resourceName, "is_enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "policies.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "policies.0.id"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.initial", "4"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.max", "5"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.capacity.0.min", "3"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.display_name", "displayName2"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.expression", "0 15 10 ? * *"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.timezone", "UTC"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.execution_schedule.0.type", "cron"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.is_enabled", "true"),
+					resource.TestCheckResourceAttr(resourceName, "policies.0.policy_type", "scheduled"),
+					resource.TestCheckResourceAttrSet(resourceName, "policies.0.time_created"),
+					resource.TestCheckResourceAttrSet(resourceName, "time_created"),
+
+					func(s *terraform.State) (err error) {
+						resId2, err = fromInstanceState(s, resourceName, "id")
+						if resId == resId2 {
+							return fmt.Errorf("Resource updated when it was supposed to be recreated.")
+						}
+						return err
+					},
+				),
+			},
+			// verify datasource
+			{
+				Config: config +
+					generateDataSourceFromRepresentationMap("oci_autoscaling_auto_scaling_configurations", "test_auto_scaling_configurations", Optional, Update, autoScalingConfigurationDataSourceRepresentation) +
+					compartmentIdVariableStr + AutoScalingConfigurationResourceDependencies +
+					generateResourceFromRepresentationMap("oci_autoscaling_auto_scaling_configuration", "test_auto_scaling_configuration", Optional, Update, autoScalingConfigurationRepresentation2),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(datasourceName, "compartment_id", compartmentId),
+					resource.TestCheckResourceAttr(datasourceName, "display_name", "displayName2"),
+
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.#", "1"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.auto_scaling_resources.#", "1"),
+					resource.TestCheckResourceAttrSet(datasourceName, "auto_scaling_configurations.0.auto_scaling_resources.0.id"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.auto_scaling_resources.0.type", "instancePool"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.compartment_id", compartmentId),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.cool_down_in_seconds", "400"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.defined_tags.%", "1"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.display_name", "displayName2"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.freeform_tags.%", "1"),
+					resource.TestCheckResourceAttrSet(datasourceName, "auto_scaling_configurations.0.id"),
+					resource.TestCheckResourceAttr(datasourceName, "auto_scaling_configurations.0.is_enabled", "true"),
+					resource.TestCheckResourceAttrSet(datasourceName, "auto_scaling_configurations.0.time_created"),
+				),
+			},
+			// verify singular datasource
+			{
+				Config: config +
+					generateDataSourceFromRepresentationMap("oci_autoscaling_auto_scaling_configuration", "test_auto_scaling_configuration", Required, Create, autoScalingConfigurationSingularDataSourceRepresentation) +
+					compartmentIdVariableStr + AutoScalingConfigurationResourceConfigForScheduledExecution,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "auto_scaling_configuration_id"),
+
+					resource.TestCheckResourceAttr(singularDatasourceName, "auto_scaling_resources.#", "1"),
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "auto_scaling_resources.0.id"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "auto_scaling_resources.0.type", "instancePool"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "compartment_id", compartmentId),
+					resource.TestCheckResourceAttr(singularDatasourceName, "cool_down_in_seconds", "400"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "defined_tags.%", "1"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "display_name", "displayName2"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "freeform_tags.%", "1"),
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "id"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "is_enabled", "true"),
+					// max_resource_count and min_resource_count are set as per the recent policy executed
+					//resource.TestCheckResourceAttrSet(singularDatasourceName, "max_resource_count"),
+					//resource.TestCheckResourceAttrSet(singularDatasourceName, "min_resource_count"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.#", "1"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.capacity.#", "1"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.capacity.0.initial", "4"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.capacity.0.max", "5"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.capacity.0.min", "3"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.display_name", "displayName2"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.execution_schedule.#", "1"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.execution_schedule.0.expression", "0 15 10 ? * *"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.execution_schedule.0.timezone", "UTC"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.execution_schedule.0.type", "cron"),
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "policies.0.id"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.is_enabled", "true"),
+					resource.TestCheckResourceAttr(singularDatasourceName, "policies.0.policy_type", "scheduled"),
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "policies.0.time_created"),
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "time_created"),
+				),
+			},
+			// remove singular datasource from previous step so that it doesn't conflict with import tests
+			{
+				Config: config + compartmentIdVariableStr + AutoScalingConfigurationResourceConfigForScheduledExecution,
+			},
+			// verify resource import
+			{
+				Config:                  config,
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{},
+				ResourceName:            resourceName,
 			},
 		},
 	})